@@ -20,16 +20,16 @@ android {
         }
     }
 
-    compileSdkVersion 27
+    compileSdkVersion 28
     defaultConfig {
         applicationId "com.android.jidian.client"
 //        compileSdkVersion rootProject.ext.android["compileSdkVersion"]
 //        buildToolsVersion rootProject.ext.android["buildToolsVersion"]
         minSdkVersion 21
-        targetSdkVersion 27
+        targetSdkVersion 28
         versionCode 1
         versionName "1.0.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
         //腾讯Bugly：使用NDK的“abiFilter”配置，设置支持的SO库架构。
         ndk {
             abiFilters 'armeabi', 'armeabi-v7a', 'arm64-v8a', 'x86', 'x86_64', 'mips', 'mips64'
