@@ -5,7 +5,7 @@ import android.graphics.Color;
 import android.graphics.Paint;
 import android.os.Handler;
 import android.os.Message;
-import android.support.v4.content.ContextCompat;
+import androidx.core.content.ContextCompat;
 import android.text.TextUtils;
 import android.view.Gravity;
 import android.view.LayoutInflater;
@@ -258,7 +258,7 @@ public class ShopRentZhima extends BaseActivity {
                             }
                         }
                     }
-                    myBaseAdapter1 = new MyBaseAdapter1(datalist_order, new int[]{R.layout.new_main_shop_order_item_top_v2, R.layout.main_shop_order_item_top2_v2, R.layout.main_shop_order_item_top_v23});
+                    myBaseAdapter1 = new MyBaseAdapter1(datalist_order, new int[]{R.layout.u6_activity_pay_create_order_item, R.layout.u6_activity_pay_create_order_item_v2, R.layout.u6_activity_pay_create_order_item_v3});
                     order_list.setAdapter(myBaseAdapter1);
                     order_list.setDividerHeight(0);
 
