@@ -1,14 +1,16 @@
 package com.android.jidian.client.adapter;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
 import com.android.jidian.client.R;
 import com.android.jidian.client.bean.AdvicesLists2Bean;
+import com.android.jidian.client.mvp.ui.activity.message.ActivityMessageViewHolder;
+import com.android.jidian.client.mvp.ui.activity.message.SystemMessageViewHolder;
 import com.android.jidian.client.util.ViewUtil;
 
 import java.util.List;
