@@ -1,6 +1,6 @@
 package com.android.jidian.client.adapter;
 
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.View;
 
 import com.android.jidian.client.bean.AdvicesLists2Bean;
