@@ -4,23 +4,13 @@ import android.app.Application;
 import android.content.Context;
 import android.content.pm.PackageInfo;
 import android.content.pm.PackageManager;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.text.TextUtils;
 
+import com.android.jidian.client.base.PermissionManager.PermissionManager;
 import com.android.jidian.client.pub.PubFunction;
 import com.android.jidian.client.util.UserInfoHelper;
-//import com.scwang.smartrefresh.header.MaterialHeader;
-//import com.scwang.smartrefresh.header.WaterDropHeader;
-//import com.scwang.smartrefresh.header.WaveSwipeHeader;
-import com.scwang.smartrefresh.layout.SmartRefreshLayout;
-import com.scwang.smartrefresh.layout.api.DefaultRefreshFooterCreator;
-import com.scwang.smartrefresh.layout.api.DefaultRefreshHeaderCreator;
-import com.scwang.smartrefresh.layout.api.RefreshFooter;
-import com.scwang.smartrefresh.layout.api.RefreshHeader;
-import com.scwang.smartrefresh.layout.api.RefreshLayout;
-import com.scwang.smartrefresh.layout.footer.ClassicsFooter;
-import com.scwang.smartrefresh.layout.header.BezierRadarHeader;
-import com.scwang.smartrefresh.layout.header.ClassicsHeader;
+
 import com.tencent.bugly.crashreport.CrashReport;
 import com.umeng.commonsdk.UMConfigure;
 import com.umeng.socialize.PlatformConfig;
@@ -41,28 +31,6 @@ public class MyApplication extends Application {
     private static Context mContext;
     /*由于骑呗后台不稳定暂时下架，暂删 public static BridgeCallback mBridgeCallback = null;
     public static String qibeidata;*/
-
-    //static 代码段可以防止内存泄露
-    static {
-        //设置全局的Header构建器
-        SmartRefreshLayout.setDefaultRefreshHeaderCreator(new DefaultRefreshHeaderCreator() {
-            @NonNull
-            @Override
-            public RefreshHeader createRefreshHeader(@NonNull Context context, @NonNull RefreshLayout layout) {
-                layout.setPrimaryColorsId(android.R.color.white);//全局设置主题颜色
-                return new ClassicsHeader(context);//.setTimeFormat(new DynamicTimeFormat("更新于 %s"));//指定为经典Header，默认是 贝塞尔雷达Header
-            }
-        });
-//        设置全局的Footer构建器
-        SmartRefreshLayout.setDefaultRefreshFooterCreator(new DefaultRefreshFooterCreator() {
-            @NonNull
-            @Override
-            public RefreshFooter createRefreshFooter(@NonNull Context context, @NonNull RefreshLayout layout) {
-                //指定为经典Footer，默认是 BallPulseFooter
-                return new ClassicsFooter(context).setDrawableSize(20).setFinishDuration(0);
-            }
-        });
-    }
 
     /**
      * 获取进程号对应的进程名
@@ -96,8 +64,23 @@ public class MyApplication extends Application {
     @Override
     public void onCreate() {
         super.onCreate();
+
         System.out.println(SHA1(getApplicationContext()));
         mContext = getApplicationContext();
+
+        //权限管理初始化
+        PermissionManager.getInstance().init(getContext());
+
+
+
+
+
+
+
+
+
+
+
         /*
          * 初始化腾讯Bugly
          */
