@@ -1,6 +1,6 @@
 package com.android.jidian.client.mvp.ui.activity.map;
 
-import android.support.v4.view.ViewPager;
+import androidx.viewpager.widget.ViewPager;
 import android.view.View;
 
 
