@@ -1,6 +1,6 @@
 package com.android.jidian.client.util;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import java.io.IOException;
 import java.nio.charset.Charset;
