@@ -3,7 +3,7 @@ package com.android.jidian.client.util;
 import android.content.Context;
 import android.content.Intent;
 import android.net.Uri;
-import android.support.v4.app.FragmentManager;
+import androidx.fragment.app.FragmentManager;
 import android.view.Gravity;
 
 import com.android.jidian.client.mvp.ui.dialog.SelectNavigationModeDialog;
