@@ -1,7 +1,7 @@
 package com.android.jidian.client.widgets;
 
 import android.app.Activity;
-import android.support.v4.view.ViewPager;
+import androidx.viewpager.widget.ViewPager;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.widget.ImageView;
